### 2024-08-19 #7115 @editorialbot
[PRE REVIEW]: BART-Survival: A Bayesian machine learning approach to survival analyses in Python
**Submitting author:** @twj8CDC (<a href="http://orcid.org/0009-0001-6265-913X">Jacob Tiegs</a>)
**Repository:** https://github.com/CDCgov/BART-Survival
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
