### 2024-08-19 #7115 @Kevin-Mattheus-Moerman
@twj8CDC- thanks for your submission to JOSS. Unfortunately, given the small size of this submission, it doesn't meet our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion. We will therefore proceed to reject this submission.

One possible alternative to JOSS is to follow [GitHub's guide](https://guides.github.com/activities/citable-code/) on how to create a permanent archive and DOI for your software. This DOI can then be used by others to cite your work.
