### 2024-08-19 #7115 @Kevin-Mattheus-Moerman
@openjournals/dev is there a way to "unreject" this submission? Or do the authors need to resubmit?
