### 2024-08-19 #7115 @twj8CDC
> @twj8CDC apologies I appear to have acted too soon. I misread the above code report (looked at the number of blank lines by accident). I will work to re-open this submission. I will however query the scope as the submission is relatively small in terms of number of lines of code. So the board will review if this submission is in scope.

@Kevin-Mattheus-Moerman
Hi Kevin, thanks starting the review process. If it helps, I understand the LOC is lower than normal, however we believe that we meet scientific contribution qualification based on the following factors:

1. There is currently no other Bayesian Additive Regression Tree (BART) for survival analysis packages ready-made in python.
2. The associated BART-Survival algorithm is complex enough that an applied practitioner would not be able to or would not want to implement the algorithm from scratch.
3. The BART-Survival algorithm is a general solution that it can be applied to most Survival modeling problems (so it is not a niche use case)

Thanks!
Jake
