### 2024-08-20 #7115 @xuanxu
There is no editorialbot command it but I've just unrejected manually
