### 2024-08-28 #7115 @Kevin-Mattheus-Moerman
@twj8CDC I am pleased to inform you this work passed the initial scope review. I have just invited an editor who can hopefully handle this submission. In the mean time can you work on the below:

- [ ] Some editors noted your README and paper are rather terse. Could you review these, and extend them where possible. I would recommend adding a brief (nearly lay person) description of the software's purpose and applications. Next I would clearly link to the documentation and testing, it may mean having to briefly reiterate some information from the documentation, but it will make the README more complete and user friendly.
- [ ] Can you try to address the potential DOI issues reported above :point_up: ? You can call `@editorialbot check references` to check them again. Also if you make changes to the paper you can call `@editorialbot generate pdf`.
