+++ requirements.txt
@@ -92,7 +92,7 @@ nbformat==5.9.2
 nbsphinx==0.9.3
 nest-asyncio==1.6.0
 nh3==0.2.15
-notebook==7.0.8
+notebook==7.2.2
 notebook_shim==0.2.3
 numba==0.58.1
 numexpr==2.9.0
