### 2024-09-04 #7115 @twj8CDC
Hi @mahfuz05062. I think L. Liu and turgeonmaxime could be potential reviewers based on their interest in Survival analysis.

How many reviewers will be selected? I can try to identify a few more if needed.

Also I am still working on updating the readme and paper per Kevins comment. I should have it updated in the next day or two.

Thanks!
