+++ README.md
@@ -13,7 +13,9 @@ Using discrete-time intervals provides a convenient approach to flexibly model t
 
 The foundation of the method is simple.  First create a sequence of time intervals, denoted as $t_j$ with ($j = {1,...,k}$), from the range of observed event times. Then for each interval $t_j$ obtain the number of observations with an event, along with the total number of observations at risk for having an event. Finally, the risk of event occurence within each interval $t_j$ can naively be derived as: 
 
-$$\begin{equation} x = 1 \end{equation}$$
+$$\begin{equation}
+x = 1 
+\end{equation}$$
 
 and the survival probability $S(t)$ at a time $q$, can be derived as:
 
