+++ README.md
@@ -81,7 +81,7 @@ The partial dependence function method is relatively simple. It involves generat
 
 The _APD_ dataset is generated so that a specific variable $x_{[I]}$ is deterministically set to a specific value for all observations while the other covariates $x_{i[O]}$ remain as observed. Each observation is then expanded over the time-intervals $1,...,j_{T_{max}}$ to create the discrete-time datasets.
 
-Mulitple _APD_ datasets can be created, each with different values of the specific variable of interest (i.e. $x_{[I]_1}$, $x_{[I]_2}$ ). The $p_{ij}$ values from each predicted dataset ($p_{[1]}$, $p_{[2]}$), can then be contrasted. 
+Mulitple _APD_ datasets can be created, each with different values of the specific variable of interest (i.e. $`x_{[I]_1}`$, $`x_{[I]_2}`$ ). The $p_{ij}$ values from each predicted dataset ($p_{[1]}$, $p_{[2]}$), can then be contrasted. 
 
 Common marginal effect estimates derived from these predicted values include:
 
