+++ README.md
@@ -89,7 +89,7 @@ Common marginal effect estimates derived from these predicted values include:
 
 $$
 \begin{equation}
-\text{Risk Diff}_{marg} = E_i S_{p_{2}}
+\text{Risk Diff}_{marg} = E_i \left[\right]
 \end{equation}
 $$
 
