+++ README.md
@@ -92,7 +92,7 @@ Common marginal effect estimates derived from these predicted values include:
 
 ```math
 \begin{equation}
-\text{Risk Diff}_{marg} = E_{i}[S_{p_{[2]}}(t_j)] - E_{i}[S_{p_{[1]}}(t_j)]
+\text{Risk Diff}_{marg} = E_{i}[S_{p_{[2]}}(t_j)] - E_{i}[S_{p_{[1]}}(t_j)] \tag{1}
 \end{equation}
 ```
 
