+++ README.md
@@ -94,7 +94,6 @@ Common marginal effect estimates derived from these predicted values include:
 \begin{equation}
 \text{Risk Diff}_{marg} = E_{i}[S_{p_{[2]}}(t_j)] - E_{i}[S_{p_{[1]}}(t_j)]
 \end{equation}
-\tag{1}
 ```
 
 - Marginal Risk Ratio at time $j$:
