+++ README.md
@@ -1,29 +1,32 @@
 ## Overview
 
-`BART-Survival` is a Python package that supports discrete-time Survival analyses using the non-parametric machine learning algorithm, Bayesian Additive Regression Trees (BART). `BART-Survival` combines the performance of the BART algorithm from the [PyMC-BART](https://www.pymc.io/projects/bart/en/latest/) library with the proper structural formatting required to complete the end-to-end Survival analysis. 
+``BART-Survival`` is a Python package that supports discrete-time Survival analyses using the non-parametric machine learning algorithm, Bayesian Additive Regression Trees (BART). ``BART-Survival`` combines the performance of the BART algorithm from the [PyMC-BART](https://www.pymc.io/projects/bart/en/latest/) library with the proper structural formatting required to complete the end-to-end Survival analysis. 
 
-`BART-Survival`'s performance is comparative to other Survival regression methods, such as Cox Proportional Hazard and AFT models, in low-complexity settings and can outperform these other models in high-complexity settings were their respective assumptions may fail (i.e. proportional hazard assumption, linearity assumptions). Additionally, the Bayesian framework provides easily accessible uncertainty intervals and nuanced interegation of a trained model.
+``BART-Survival``'s performance is comparative to other Survival regression methods, such as Cox Proportional Hazard and AFT models, in low-complexity settings and can outperform these other models in high-complexity settings were their respective assumptions may fail (i.e. proportional hazard assumption, linearity assumptions). Additionally, the Bayesian framework provides easily accessible uncertainty intervals and capabilities to extensively interrogate trained model.
 
-The `BART-Survival` library provides a simple API for completing standard Survival analysis, as well as allowing exposure to the underlying PyMC code providing accessibility to extend the `BART-Survival` library when necessary.
+The ``BART-Survival`` library provides a simple API for completing standard Survival analysis, as well as allowing exposure to the underlying PyMC code providing accessibility to extend the ``BART-Survival`` library when necessary.
 
-This repository contains the source code and documentation for the BART-Survival package as well as [example](https://github.com/CDCgov/BART-Survival/blob/main/examples/) notebooks. 
+This repository contains the source code and documentation for the `BART-Survival` package as well as [example](https://github.com/CDCgov/`BART-Survival`/blob/main/examples/) notebooks. 
 
-`BART-Survival` can be installed directly from PyPi 
+``BART-Survival`` can be installed directly from PyPi 
 ```python
-pip install bart-survival==0.1.1
+pip install `BART-Survival`==0.1.1
 ```
-Or as downloaded from the repository. 
+Or as accessed from this repository. 
 
 
 
 ### Background
 
-Survival analysis methods are statistical methods used to describe the risk of an event occurrence over a period of time. The BART-Survival package provides a discrete-time survival method which aims to model survival as a function of the cumulative risk of event occurence over the series of discrete time intervals. 
+Survival analysis methods are statistical methods used to describe the risk of an event occurrence over a period of time. The `BART-Survival` package provides a discrete-time Survival method which aims to model Survival as a function of the cumulative risk of event occurrences over the series of discrete time intervals. 
 
-Using discrete-time intervals provides a convenient approach to flexibly model the latent probability of event as a non-parametric function of the distinct time interval and a set of observation covariates. The latent probabilities can then be used for deriving survival probability or other estimates.
[…]
