### 2024-09-09 184f04f
draft paper
