+++ paper.md
@@ -41,6 +41,131 @@ More recently, non-parametric machine learning approaches have been introduced t
 Currently, the only BART survival algorithm readily available exists as part of the `BART` R package, which contains a library of various BART-based approaches in addition to a BART survival analysis application [@sparapani2016; @sparapani2021]. BART-Survival package described here combines the survival analysis approach outlined in the `BART` R package with the foundational Python-based probabilistic programming language library, `PyMC`, and the accompanying BART algorithm from the `PyMC-BART` library.
 Our aim in developing `BART-Survival` is to provide accessibility to the BART survival algorithm within the Python programing language. This contribution is beneficial for analysts when Python is the preferred programming language, the analytic workflow is Python-based, or when the R language is unavailable for analyses. Additionally, `BART-Survival` package abstracts away the complexities of the `PyMC` and `PyMC-BART` libraries through use of a pre-specified core model and generalized functionality that can accommodate analyses across various survival settings. The  `BART-Survival` package is intended for public health and clinical professionals and students who are looking for non-parametric alternatives to traditional (semi-)parametric survival analysis, especially for use in large, complex healthcare data and machine learning applications. 
 
+# Methods
+
+The following sections provides details on the methods employed by the BART-Survival library, focusing specifically on the discrete-time Survival algorithm and the produced estimates. For review of the BART algorithm used we refer to PyMC-BART publication [@quiroga2023].
+
+## Background
+The `BART-Survival` package provides a discrete-time Survival method which aims to model Survival as a function of the cumulative risk of event occurrences over the series of discrete time intervals. 
+
+Using discrete-time intervals provides a convenient approach that allows flexible modeling of the latent probability of an event as a non-parametric function of the distinct time interval and a set of observation covariates. The latent probabilities can then be used for deriving Survival probability or other estimates.
+
+
+The foundation of the method is simple.  
+1. Create a sequence of time intervals, denoted as $t_j$ with ($j = {1,...,k}$), from the range of observed event times. 
+2. Then for each interval $t_j$ obtain the number of observations with an event, along with the total number of observations at risk for having an event. 
+3. Finally, the risk of event occurrence within each interval $t_j$ can naively be derived as: 
+
+```math
+\begin{equation}
+P_{t_j} = \frac {\text{n events}_{t_j}} {\text{n at risk}_{t_j}}
+\end{equation}
+```
+
+and the Survival probability $S(t)$ at a time $q$, can be derived as:
+
+```math
+\begin{equation}
+S(t_q) = \prod_{j=1}^{q} (1-P_{t_j}) 
+\end{equation}
+```
+
+`BART-Survival` builds off this simple foundation by replacing $P_t$ with a probability risk estimate, $p_{t_j|x_i}$ yielded from the BART regression model. The predicted values $p_{t_j|x_i}$ are generated for each observation, at each time interval from the set $j$. Downstream targets can be further derived from these predicted values with observation-level Survival derived as:
+
+```math
+\begin{equation}
+S(t_q|x_i) = \prod_{j=1}^{q} (1-p_{t_j|x_i})
+\end{equation}
+```
+Statistical estimands can also be estimated using the predicted $p_{t_j|x_i}$ through evaluation of marginal functions of the predicted values. 
+
+## Data Preparation  
[…]
