### 2024-09-10 cc5fb25
updated paper, readme, bib
