+++ paper.bib
@@ -24,6 +24,7 @@
   month = aug,
   number = {arXiv:2206.03619},
   eprint = {2206.03619},
+  doi = {10.48550/arXiv.2206.03619},
   primaryclass = {stat},
   publisher = {arXiv},
   urldate = {2024-03-11},
@@ -185,11 +186,12 @@
 }
 
 @misc{saha2019,
-  title = {Variable {{Selection}} with {{Random Survival Forest}} and {{Bayesian Additive Regression Tree}} for {{Survival Data}}},
+  title =  {{Selection}} with {{Random Survival Forest}} and {{Bayesian Additive Regression Tree}} for {{Survival Data}}},
   author = {Saha, Satabdi and Ryu, Duchwan and Ebrahimi, Nader},
   year = {2019},
   month = nov,
   number = {arXiv:1910.02160},
+  doi = {10.48550/arXiv.1910.02160},
   eprint = {1910.02160},
   primaryclass = {stat},
   publisher = {arXiv},
