+++ paper.bib
@@ -185,7 +185,7 @@
   file = {C:\Users\twj8\Zotero\storage\S64BXRLN\Sparapani et al. - 2021 - Nonparametric Machine Learning and Efficient Compu.pdf}
 }
 
-@misc{saha2019,
+@article{saha2019,
   title =  {{Selection}} with {{Random Survival Forest}} and {{Bayesian Additive Regression Tree}} for {{Survival Data}}},
   author = {Saha, Satabdi and Ryu, Duchwan and Ebrahimi, Nader},
   year = {2019},
