+++ paper.bib
@@ -186,7 +186,7 @@
 }
 
 @article{saha2019,
-  title =  {{Selection}} with {{Random Survival Forest}} and {{Bayesian Additive Regression Tree}} for {{Survival Data}}},
+  title =  {{{Selection}} with {{Random Survival Forest}} and {{Bayesian Additive Regression Tree}} for {{Survival Data}}},
   author = {Saha, Satabdi and Ryu, Duchwan and Ebrahimi, Nader},
   year = {2019},
   month = nov,
