### 2024-09-11 #7213 @twj8CDC
@WeakCha
Thank you for the initial round of edits. I will work on updating based on these comments.

As far as point 10. We did do an extensive validation analysis for our method with simulations and real data, in which we compared our algorithm to the R based algorithm. However, I was under  the impression that JOSS did not necessarily want this type of analysis in the JOSS papers. We were planning to put together the validation study in a separate paper.

I could provide an example comparing our algorithm and the R algorithm using a simulated dataset if you think that would be sufficient. Or if you have other thoughts?
