### 2024-09-30 77b703c
updated paper, examples
