+++ paper.md
@@ -52,23 +52,99 @@ In combination with a structural configuration of the data, the discrete-time al
 
 The foundation of the method is simple.  
 
-1. Create a sequence of time intervals, denoted as $t_j$ with ($j = {1,...,k}$), from the range of observed event times. 
 
-2. Then for each interval $t_j$ obtain the number of observations with an event, along with the total number of observations at risk for having an event. 
+
+1. Starting with a simple event-time dataset, create a sequence of time intervals that represent the unique discrete-time intervals observed in the data. Each intervals are represented as a $t_j$, where $j = {1,...,k}$ and $k$ is the length of set of unique observed times.
+
+
+
+    For example if the observed event-time data is:
+
+    $$
+    \begin{matrix} 
+    \text{event status} &\text{event time} \\
+    --- &---\\
+        1 & 1 \\
+        0 & 2 \\
+        1 & 2 \\
+        1 & 4 \\
+        1 & 4 \\
+        1 & 5 \\
+    \end{matrix}
+    $$
+
+    Then the set of unique observed times is $[1,2,4,5]$ with $k = 4$ indices and the corresponding $t_j$ intervals are:
+
+    $$
+    \begin{matrix} 
+    \text{index} &\text{interval} \\
+    --- &---\\
+        t_1 & [0,1) \\
+        t_2 & [1,2) \\
+        t_3 & [2,4) \\
+        t_4 & [4,5) \\
+    \end{matrix}
+    $$
+    
+    of note:
+    - "$[$" is exclusive and "$)$" is inclusive times in the interval
+    - for **event status**: $(1 = \text{event}, 0 = \text{censor})$
+
+
+2. For each interval tally the number of observations with an event, the number of observation censored and the total number of observations eligible to have an event at the start of each interval $t_j$.
+
+    Continuing the above example the corresponding frequencies for each interval are:
+
+    $$
+    \begin{matrix} 
+    \text{index} &\text{interval} &\text{event} &\text{censor} & \text{at risk}\\
+    --- &---&---&---&---\\
+        t_1 & [0,1) & 1  &   & 6 \\
+        t_2 & [1,2) &  1 & 1 & 5 \\
+        t_3 & [2,4) &  2 &   & 3 \\
+        t_4 & [4,5) &  1 &   & 1 \\
+    \end{matrix}
+    $$
+
 
 3. Finally, the risk of event occurrence within each interval $t_j$ can naively be derived as: 
 
-$$
-P_{t_j} = \frac {\text{n events}_{t_j}} {\text{n at risk}_{t_j}}
-$$
+    $$
+    P_{t_j} = \frac {\text{n events}_{t_j}} {\text{n at risk}_{t_j}}
+    $$
 
-and the Survival probability $S(t)$ at a time $q$, can be derived as:
+    and the Survival probability at a time-index $q$, can be derived as:
 
-$$
-S(t_q) = \prod_{j=1}^{q} (1-P_{t_j}) 
-$$
+    $$
+    S(t_q) = \prod_{j=1}^{q} (1-P_{t_j}) 
+    $$
 
-`BART-Survival` builds off this simple foundation by replacing $P_t$ with a probability risk estimate, $p_{t_j|x_i}$ yielded from the BART regression model. The predicted values $p_{t_j|x_i}$ are generated for each observation, at each time interval from the set $j$. Downstream targets can then be derived from these predicted values. For example the Survival probability curve for a single observation can be derived as:
+    Applied to our example the risks of event $P_{t_j}$ are:
+
+    $$
+    \begin{matrix} 
[…]
