### 2024-10-01 #7213 @mahfuz05062
@WeakCha and @twj8CDC, I don't think an extensive analysis is necessary here but I like the idea mentnioned by @twj8CDC "comparing our algorithm and the R algorithm using a simulated dataset". If you can provide something like this, it should be sufficient.
