+++ paper.md
@@ -43,18 +43,17 @@ Our aim in developing `BART-Survival` is to provide accessibility to the BART su
 
 # Methods
 
-The following sections provides details on the methods employed by the BART-Survival library, focusing specifically on the discrete-time Survival algorithm used. For review of the BART algorithm we refer to associated PyMC-BART publication [@quiroga2023].
+The following sections provides details on the methods employed by the `BART-Survival` library, focusing specifically on the discrete-time Survival algorithm used. For review of the BART algorithm we refer to associated `PyMC-BART` publication [@quiroga2023].
 
 ## Background
-The `BART-Survival` package provides a discrete-time Survival method which aims to model Survival as a function of a series of probabilities (indicating risk of event) that can be determined from the sequence of discrete-time intervals examined. 
-In combination with a structural configuration of the data, the discrete-time algorithm allows for flexible modeling of the risk probabilities as a non-parametric function of  time and observed covariates. The series of probability risks can then be used in deriving Survival probabilities, along with other estimates of interest.
+The `BART-Survival` package provides a discrete-time Survival method which aims to model Survival as a function of a series of probabilities that indicate the probability of event occurrence at each discrete time. For clarity, the probability of event occurrence at each discrete time will be referred to as the risk probability.
+
+In combination with a structural configuration of the data, the discrete-time algorithm allows for flexible modeling of the risk probabilities as a non-parametric function of  time and observed covariates. The series of probability risks can then be used to derive the Survival probabilities, along with other estimates of interest.
 
 
 The foundation of the method is simple.  
 
-
-
-1. Starting with a simple event-time dataset, create a sequence of time intervals that represent the unique discrete-time intervals observed in the data. Each intervals are represented as a $t_j$, where $j = {1,...,k}$ and $k$ is the length of set of unique observed times.
+1. Starting with a simple event-time dataset, create a sequence of time intervals that represent the unique discrete-time intervals observed in the data. Each interval is represented as a $t_j$, where $j = {1,...,k}$ and $k$ is the length of the set of unique observed times.
 
 
 
@@ -86,12 +85,13 @@ The foundation of the method is simple.
     \end{matrix}
     $$
     
-    of note:
-    - "$[$" is exclusive and "$)$" is inclusive times in the interval
-    - for **event status**: $(1 = \text{event}, 0 = \text{censor})$
+    Where the intervals are denoted with the properties, "$[$" indicating exclusive times and "$)$" indicating inclusive times in an interval. Additionally, the **event status** column is defined as $(1 = \text{event}, 0 = \text{censor})$
 
 
-2. For each interval tally the number of observations with an event, the number of observation censored and the total number of observations eligible to have an event at the start of each interval $t_j$.
+2. Then with the constructed time intervals $t_1,...,t_k$, tally the following for each interval:
+    - the number of observations with an event
+    - the number of observation censored 
+    - the total number of observations eligible to have an event  (at risk) at the start of the interval
 
     Continuing the above example the corresponding frequencies for each interval are:
 
@@ -107,16 +107,16 @@ The foundation of the method is simple.
     $$
 
 
-3. Finally, the risk of event occurrence within each interval $t_j$ can naively be derived as: 
+3. Finally, the risk of event occurrence within each interval $t_j$ can simply be derived as: 
 
     $$
[…]
