### 2024-10-22 #7213 @turgeonmaxime
@twj8CDC Thanks for your updates to the paper. I think one crucial point that is still lacking from the manuscript is what role `BART-Survival` fulfills that `PyMC-BART` doesn't. Based on my reading of the manuscript but also pre-review discussions here on Github, generating the datasets required for the analysis is a non-trivial task. You should write this explicitly in the text. It is the main message.

Similarly, I think you did a good job explaining the need for `BART-Survival` when a similar R package already exists. And that's also a crucial requirement, so thanks for that.

Beyond that, I have a few more comments, some general and some more targeted:

- In general, it is recommended to use punctuation with equations, as they are generally part of a sentence. For example, on line 81, the equation should end with a comma, and the equation on line 82 should end with a period. Moreover, if you add extra context after an equation (e.g. line 185), the equation above should end with a comma and the sentence that follows should start with a small letter (i.e. "where" instead of "Where").
- Related to the comment above, it would be awkward to try and do the same with the tables that are part of your sentences (e.g. lines 69 and 71), so I think it's fine to drop the punctation in those cases. I would therefore recommend rephrasing line 72 so that you get a complete sentence. For example, "In the table above, the intervals are..."
- I'm not sure why "Survival" appears multiple times with a capital S. Please review and fix as needed.
- What you describe as the "simple setting" in the "Background" subsection is essentially the Kaplan-Meier non-parametric estimator. I would suggest you add that clarification and maybe also a reference to your favourite textbook on survival analysis.
- L124: More generally, the dataset TAD can be used for **binary regression**. Probit regression is just one type of binary regression.
- L147: PAD length should be 48 (which was also mentioned by @WeakCha).
- L161: Typo, "the ~of the~ probability"
- L201: Missing a period at the end of the sentence.
- L235: The sentence should be reviewed, I think it should be "These expectations can be further used to make comparisons ~can be made~ between".
- L265: what are the `weights` and `coords` parameters?
- L268: I think it should be a small s in "The p,S arrays..."
- L294: There is a typo in the URL, hence why the link doesn't work.

I'll have a look at the repo itself and report back.
