### 2024-10-22 #7213 @turgeonmaxime
@twj8CDC One more thing: could you add a short description of the `rossi` dataset that you used for the demonstration. Just a sentence or two would be helpful (especially since the `lifelines` documentation doesn't even mention what the "survival" event is).
