### 2024-10-29 #7213 @turgeonmaxime
@twj8CDC I had a look at the repo. I was able to install the package and run the example code without any issue. So from my perspective, this part meets the requirements, and I've updated my checklist above to reflect that.

Again from my perspective, the only thing missing is to address the last comments about the manuscript.
