+++ paper.bib
@@ -50,6 +50,10 @@
   file = {C:\Users\twj8\Zotero\storage\3R4NAUEW\Patilea and Van Keilegom - 2020 - A general approach for cure models in survival ana.pdf}
 }
 
+
+
+
+
 @article{logan2019,
   title = {Decision Making and Uncertainty Quantification for Individualized Treatments Using {{Bayesian Additive Regression Trees}}},
   author = {Logan, Brent R and Sparapani, Rodney and McCulloch, Robert E and Laud, Purushottam W},
@@ -150,6 +154,27 @@
   abstract = {Bayesian additive regression trees (BART) provide a framework for flexible nonparametric modeling of relationships of covariates to outcomes. Recently, BART models have been shown to provide excellent predictive performance, for both continuous and binary outcomes, and exceeding that of its competitors. Software is also readily available for such outcomes. In this article, we introduce modeling that extends the usefulness of BART in medical applications by addressing needs arising in survival analysis. Simulation studies of one-sample and two-sample scenarios, in comparison with long-standing traditional methods, establish face validity of the new approach. We then demonstrate the model's ability to accommodate data from complex regression models with a simulation study of a nonproportional hazards scenario with crossing survival functions and survival function estimation in a scenario where hazards are multiplicatively modified by a highly nonlinear function of the covariates. Using data from a recently published study of patients undergoing hematopoietic stem cell transplantation, we illustrate the use and some advantages of the proposed method in medical investigations. Copyright {\copyright} 2016 John Wiley \& Sons, Ltd.},
   langid = {english},
   file = {C:\Users\twj8\Zotero\storage\N4V845N4\Sparapani et al. - 2016 - Nonparametric survival analysis using Bayesian Add.pdf}
+}
+
+@article{rossi1980,
+  title = {Money, Work, and Crime: Some Experimental Results},
+  author = {Rossi, P.H., R.A. Berk, and K.J. Lenihan},
+  year = {1980},
+  journal = {New York: Academic Press}
+}
+
+@article{stel2011,
+  title = {Survival analysis I: the Kaplan-Meier method},
+  author = {Stel Vianda S., Dekker Friedo W., Tripepi Giovanni, Zoccali Carmine, Jager Kitty J.},
+  year = {2011},
+  month = jun,
+  journal = {Nephron Clin Pract},
+  volume = {119},
+  number = {1},
+  issn = {1660-2110},
+  doi = {10.1159/000324758},
+  urldate = {2024-10-31},
+  langid = {english}  
 }
 
 @article{chakraborty2022,
