+++ README.md
@@ -18,127 +18,6 @@ Or as accessed from this repository.
 [BART-Survival API](https://cdcgov.github.io/BART-Survival/build/html/index.html) 
 
 
-
-### Background
-
-Survival analysis methods are statistical methods used to describe the risk of an event occurrence over a period of time. The `BART-Survival` package provides a discrete-time Survival method which aims to model Survival as a function of the cumulative risk of event occurrences over the series of discrete time intervals. 
-
-Using discrete-time intervals provides a convenient approach to flexibly model the latent probability of event as a non-parametric function of the distinct time interval and a set of observation covariates. The latent probabilities can then be used for deriving Survival probability or other estimates.
-
-
-The foundation of the method is simple.  
-1. Create a sequence of time intervals, denoted as $t_j$ with ($j = {1,...,k}$), from the range of observed event times. 
-2. Then for each interval $t_j$ obtain the number of observations with an event, along with the total number of observations at risk for having an event. 
-3. Finally, the risk of event occurrence within each interval $t_j$ can naively be derived as: 
-
-```math
-\begin{equation}
-P_{t_j} = \frac {\text{n events}_{t_j}} {\text{n at risk}_{t_j}}
-\end{equation}
-```
-
-and the Survival probability $S(t)$ at a time $q$, can be derived as:
-
-```math
-\begin{equation}
-S(t_q) = \prod_{j=1}^{q} (1-P_{t_j}) 
-\end{equation}
-```
-
-`BART-Survival` builds off this simple foundation by replacing $P_t$ with a probability risk estimate, $p_{t_j|x_i}$ yielded from the BART regression model. The predicted values $p_{t_j|x_i}$ are generated for each observation, at each time interval from the set $j$. Downstream targets can be further derived from these predicted values with observation-level Survival derived as:
-
-```math
-\begin{equation}
-S(t_q|x_i) = \prod_{j=1}^{q} (1-p_{t_j|x_i})
-\end{equation}
-```
-Statistical estimands can also be estimated using the predicted $p_{t_j|x_i}$ through evaluation of marginal functions of the predicted values. 
-
-To properly model $p_{t|x}$ the data requires an transformation from a standard Survival dataset to a _augmented_ dataset. Survival data is typically given as a paired (**event status**, **event time**) outcome, along with a set of covariates for each observation. In this setup **event status** is typically a binary variable (1=event; 0=censored) and **event time** is some continuous representation of time.
-
-The _augmented_ dataset transforms the generic dataset from a single, paired (**event status**, **event time**) outcome per observation, to a sequence of single (**event status**) outcomes over the series of discrete-time intervals, up to the given **event time** for the observation. 
-
-For example if the unique set of a dataset's **event times** is **{4,6,7,8,12,14}**, and a single observation's paired outcome is (**event status** = 1, **event times** = 12), then the single observation will be represented in the _augmented_ dataset as the sequence of rows:
-
-```math
-\begin{matrix} 
-\text{event status} &\text{time} \\
- --- &---\\
-    0 & 4\\
-    0 & 6 \\
-    0 & 7 \\
-    0 & 8 \\
-    1 & 12 \\
-\end{matrix}
-```
-
[…]
