### 2024-11-01 #7213 @turgeonmaxime
@twj8CDC Thanks, this looks good to me! Although there is still a typo on line 157, it should say "(8 * 6) = 48".

@mahfuz05062 I've completed the review and updated the checklist. I recommend the manuscript be accepted by JOSS (pending the correction to the typo above).
