### 2024-11-04 #7213 @twj8CDC
@WeakCha Hi, thank you for the feedback. The bracketed subscript is meant to identify the covariate selected to be deterministically set in each PDAD. I adjusted this section to be more explicit regarding what this notation means. Let me know if it is now more clear! Thanks
