### 2024-11-12 4a3e864
add doi to readme
