### 2024-11-12 #7213 @twj8CDC
[x] Double check authors and affiliations (including ORCIDs)

[x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

[x] Archive the release on Zenodo/figshare/etc and post the DOI here.
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14105188.svg)](https://doi.org/10.5281/zenodo.14105188)

[x] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
[x] Make sure that the license listed for the archive is the same as the software license.

@mahfuz05062 The above tasks have been completed. Please let me know if you have any questions/concerns. Thanks!
