### 2024-12-06 #7213 @mahfuz05062
> hi @mahfuz05062 just wanted to follow-up on this to make sure there is no remaining tasks from my end required to complete the submission. Thanks!

Thanks for the follow-up. I have actually gone through some parts of your paper but it is quite large compared to our usual submissions so I will need some more time to complete my review. However, I plan to finish it this weekend.

Once I do that, I will have some comments for you, which will be about language, typos, etc., and won't be about the content as the reviewers are already happy about it.
