### 2024-12-09 #7213 @mahfuz05062
Hi @twj8CDC, I went over the articles. Following are some of the suggestions:

In abstract, can you simplify this "BART-Survival combines the performance of the BART algorithm with the complementary data and model structural formatting required to complete the survival analyses." For exmaple, if the word "structural" is not required here, you can just remove it.

Line 58, "it is expected users" should be "it is expected that users"
Missing comma on line 60, line 98, 136, 137, 148, in other places.

I would suggest go over one more time and make any grammatical corrections like this is needed.
