### 2024-12-09 641a858
paper check
