### 2024-12-23 #7213 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
