### 2025-01-08 #7213 @twj8CDC
Hi @mahfuz05062,

Do you know how long it typically take for the publication process to be finalized?
There are no final steps required from my end correct?

Thanks
Jake
