### 2025-01-09 #7213 @Kevin-Mattheus-Moerman
- [ ] In your affiliations please spell out USA as United States or United States of America.
- [ ] `github` should read `Github`
- [ ] Our guidelines state `The paper should be between 250-1000 words`. At ~5000 words, the paper is therefore currently too long. I suggest you remove the `Methods` and `Demonstration` sections fully and instead incorporate (if you haven't already) such material in your project documentation.
