### 2025-01-13 d2bc7bf
update documentation (moved methods)
