+++ paper_full.md
@@ -1,0 +1,664 @@
+
+---
+title: 'BART-Survival: A Bayesian machine learning approach to survival analyses in Python'
+tags:
+  - Python
+  - Bayesian 
+  - Machine Learning
+  - Survival
+  - Time to Event
+authors:
+  - name: Jacob Tiegs
+    orcid: 0009-0001-6265-913X
+    affiliation: "1, 2" 
+  - name: Julia Raykin
+    orcid: 0009-0006-1840-6991
+    corresponding: true 
+    affiliation: 1
+  - name: Ilia Rochlin
+    orcid: 0000-0001-7680-6965
+    corresponding: true 
+    affiliation: 1
+affiliations:
+ - name: Inform and Disseminate Division, Office of Public Health Data, Surveillance, and Technology, Centers for Disease Control and Prevention, Atlanta, GA, USA
+   index: 1
+ - name: Metas Solutions, Atlanta, Georgia
+   index: 2
+
+date: 08 August 2024
+bibliography: paper.bib
+
+---
+
+# Summary
+
+`BART-Survival` is a Python package that allows time-to-event (survival) analyses in discrete-time using the non-parametric machine learning algorithm, Bayesian Additive Regression Trees (BART). `BART-Survival` combines the performance of the BART algorithm with the complementary data and model formatting required to complete the survival analyses. The library contains a convenient application programming interface (API) that allows a simple approach when using the library for survival analyses, while maintaining capabilities for added complexity when desired. The package is intended for analysts exploring use of flexible non-parametric alternatives to traditional (semi-)parametric survival analyses. 
+
+# Statement of need
+
+Survival analyses are a cornerstone of public health and clinical research in such diverse fields as cancer, cardiovascular disease, and infectious diseases [@altman1998; @bradburn2003]. Traditional parametric and semi-parametric statistical methods, such as the Cox proportional hazards model, are commonly employed for survival analyses [@cox1972]. However, these methods have several limitations, particularly when applied to complex data. One major issue is the need for restrictive assumptions, such as proportional hazards and predefined functional forms, which may not hold true in complex, real-world healthcare data [@ishwaran2008; @harrell2015]. Additionally, these methods often struggle with high-dimensional datasets, leading to problems with over-fitting, multi-collinearity, and dealing with complex interactions [@ishwaran2008; @joffe2013].
+
+More recently, non-parametric machine learning approaches have been introduced to address the limitations of the traditional methods [@ishwaran2008; @harrell2015]. BART is a one such approach that has demonstrated exceptional performance across a variety of analytic settings, typically outperforming the traditional methods in terms of predictive accuracy. BART's performance is linked to it ability to flexibly model complex non-linear and variable interactions within the data, while being inherently regularized to reduce issues of over-fitting. The BART method is fully non-parametric and can adaptively model data complexities without prior knowledge or specification of a particular functional form. Finally, the method is generally accepted as being a user-friendly machine learning approach, as it typically requires minimal hyperparameter tuning and the outcomes can be easier to interpret than those produced by other similar methods [@chipman2010; @sparapani2016;@sparapani2021]. 
+
[…]
