+++ paper.md
@@ -20,7 +20,7 @@ authors:
     corresponding: true 
     affiliation: 1
 affiliations:
- - name: Inform and Disseminate Division, Office of Public Health Data, Surveillance, and Technology, Centers for Disease Control and Prevention, Atlanta, GA, USA
+ - name: Inform and Disseminate Division, Office of Public Health Data, Surveillance, and Technology, Centers for Disease Control and Prevention, Atlanta, GA, United States of America
    index: 1
  - name: Metas Solutions, Atlanta, Georgia
    index: 2
@@ -49,602 +49,6 @@ More specifically, the `BART-Survival` library abstracts away the complexities o
 
 In summary, the `BART-Survival` package provides a simple and accessible approach to implementing the BART survival algorithm. The provided approach can be beneficial for users who are looking for non-parametric alternatives to traditional (semi-)parametric survival analysis. The BART survival algorithm can be especially useful in large, complex healthcare data, where machine learning methods can demonstrate improved performance over the traditional methods. 
 
-# Methods
-
-The following sections provides details on the methods employed by the `BART-Survival` library, focusing specifically on the discrete-time survival algorithm used. For review of the BART algorithm we refer to associated `PyMC-BART` publication [@quiroga2023].
-
-## Background
-The `BART-Survival` package provides a discrete-time survival method which aims to model survival as a function of a series of probabilities that indicate the probability of event occurrence at each discrete time. For clarity, the probability of event occurrence at each discrete time will be referred to as the risk probability.
-
-In combination with a structural configuration of the data, the discrete-time algorithm allows for flexible modeling of the risk probabilities as a non-parametric function of time and observed covariates. The series of probability risks can then be used to derive the survival probabilities, along with other estimates of interest.
-
-
-The foundation of the method is simple and is based off the well-defined Kaplan-Meier Product-Limit estimator. While a full review of the Kaplan-Meier method can be found elsewhere [@stel2011], the following example demonstrates the fundamental concepts of discrete-time survival analysis that motivates its application within the `BART-Survival` method.
-
-1. Starting with a simple event-time dataset, create a sequence of time intervals that represent the unique discrete-time intervals observed in the data. Each interval is represented as a $t_j$, where $j = {1,...,k}$ and $k$ is the length of the set of unique observed times.
-
-
-
-    For example if the observed event-time data is:
-
-    $$
-    \begin{matrix} 
-    \text{event status} &\text{event time} \\
-    --- &---\\
-        1 & 1 \\
-        0 & 2 \\
-        1 & 2 \\
-        1 & 4 \\
-        1 & 4 \\
-        1 & 5 \\
-    \end{matrix}
-    $$
-
-    Then the set of unique observed times is $[1,2,4,5]$ with $k = 4$ indices and the corresponding $t_j$ intervals are:
-
-    $$
-    \begin{matrix} 
-    \text{index} &\text{interval} \\
-    --- &---\\
-        t_1 & (0,1] \\
-        t_2 & (1,2] \\
-        t_3 & (2,4] \\
-        t_4 & (4,5] \\
-    \end{matrix}
-    $$
-    
[…]
