### 2025-01-15 #7213 @Kevin-Mattheus-Moerman
@twj8CDC I reviewed the new paper and the formatting/length looks much better thanks. Below are some minor final points.

- [ ] Please also add `United States of America` for the second affiliation.
- [ ] Check `A extended review of method` which should read `An extended review of the methods` (or feel free to edit/fix it differently as you see fit)
- [ ] Where you say `While the manuscript for this work is currently under development..` I recommend that you upload the draft, if possible, to a pre-print server (such as the ArXiv) such that you can cite it in this JOSS paper. This is not a requirement but it is highly recommended if possible.
