### 2025-01-15 #7213 @twj8CDC
@Kevin-Mattheus-Moerman sounds good! I have resolved the first two points.

I will discuss with my team regarding the last point. I think you are correct that this would be better. I am currently working on completing that paper and could probably upload a preprint draft to arxiv by early next week.

I will let you know the plan in a couple of days if that is okay with you?
